Remove AccessibleRobot and refactor element mechanism
