+++ DropBear/Robot/Robot+Assert.swift
@@ -3,35 +3,38 @@ import XCTest
 extension Robot {
     @discardableResult
     public func assert(
-        identifier: String,
+        _ element: Element,
         in hierarchy: [XCUIElement.ElementType] = [.any],
         _ assertion: ElementAssertion,
         file: StaticString = #file, line: UInt = #line
         ) -> Self
     {
-        return assert(identifier: identifier, in: hierarchy, [assertion], file: file, line: line)
+        return assert(element, in: hierarchy, [assertion], file: file, line: line)
     }
 
     @discardableResult
     public func assert(
-        identifier: String,
+        _ element: Element,
         in hierarchy: [XCUIElement.ElementType] = [.any],
         _ assertion: ElementAssertion, _ rest: ElementAssertion...,
         file: StaticString = #file, line: UInt = #line
         ) -> Self
     {
-        return assert(identifier: identifier, in: hierarchy, [assertion] + rest, file: file, line: line)
+        return assert(element, in: hierarchy, [assertion] + rest, file: file, line: line)
     }
 
-    func assert(
-        identifier: String,
+    @discardableResult
+    public func assert(
+        _ element: Element,
         in hierarchy: [XCUIElement.ElementType] = [.any],
         _ assertions: [ElementAssertion],
         file: StaticString = #file, line: UInt = #line
         ) -> Self
     {
-        let element = self.element(identifier: identifier, in: hierarchy, file: file, line: line)
-        _ = element.assert(assertions, file: file, line: line)
+        element
+            .element(in: app, hierarchy: hierarchy, file: file, line: line)
+            .assert(assertions, file: file, line: line)
+
         return self
     }
 }
