+++ DropBear/Robot/Robot.swift
@@ -2,7 +2,8 @@ import XCTest
 
 /// A `Robot` represents the actions that can be performed on a given screen
 public protocol Robot {
-    associatedtype Container: ContainerType
+    associatedtype Container: ContainerType = Containers.None
+    associatedtype Element: RawRepresentable = String where Element.RawValue == String
 
     var app: XCUIApplication { get }
 
@@ -18,13 +19,27 @@ extension Robot {
 }
 
 
-/// A pre-made base class you can use to avoid the `Robot` boilerplate code
-/// It does not conform to `Robot` so consumers are still required to conform and
-/// provide a value for `Container` 
+/// A pre-made base class you can use to avoid some `Robot` boilerplate code
+/// It does not conform to `Robot` so consumers are still required to do so
 open class RobotBase {
     public let app: XCUIApplication
 
     public required init(app: XCUIApplication) {
         self.app = app
+    }
+}
+
+extension RawRepresentable where RawValue == String {
+    public func element(in application: XCUIApplication, hierarchy: [XCUIElement.ElementType], file: StaticString, line: UInt) -> XCUIElement {
+        return application.element(identifier: rawValue, in: hierarchy, file: file, line: line)
+    }
+}
+
+extension String: RawRepresentable {
+    public init?(rawValue: String) { self = rawValue }
+    public var rawValue: String { return self }
+
+    public func element(in application: XCUIApplication, hierarchy: [XCUIElement.ElementType], file: StaticString, line: UInt) -> XCUIElement {
+        return application.element(identifier: self, in: hierarchy, file: file, line: line)
     }
 }
